python-assessment-task/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
cramjam==2.9.0
fastparquet==2024.5.0
fsspec==2024.9.0
idna==3.10
load-dotenv==0.1.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
